+++ lib/ex_nylas/notetaker/build.ex
@@ -8,6 +8,7 @@ defmodule ExNylas.Notetaker.Build do
   use TypedEctoSchema
   import Ecto.Changeset
   alias ExNylas.Schema.Util
+  alias ExNylas.Notetaker.CustomSettings
 
   @derive {Jason.Encoder, only: [:meeting_link, :join_time, :name, :meeting_settings, :rules]}
   @primary_key false
@@ -18,11 +19,16 @@ defmodule ExNylas.Notetaker.Build do
     field(:name, :string)
 
     embeds_one :meeting_settings, MeetingSettings, primary_key: false do
-      @derive {Jason.Encoder, only: [:video_recording, :audio_recording, :transcription]}
+      @derive {Jason.Encoder, only: [:video_recording, :audio_recording, :transcription, :action_items, :summary, :action_items_settings, :summary_settings]}
 
       field(:video_recording, :boolean)
       field(:audio_recording, :boolean)
       field(:transcription, :boolean)
+      field(:action_items, :boolean)
+      field(:summary, :boolean)
+
+      embeds_one :action_items_settings, CustomSettings
+      embeds_one :summary_settings, CustomSettings
     end
 
     embeds_one :rules, Rules, primary_key: false do
@@ -44,7 +50,14 @@ defmodule ExNylas.Notetaker.Build do
     struct
     |> cast(params, [:name, :join_time, :meeting_link])
     |> validate_required([:meeting_link])
-    |> cast_embed(:meeting_settings, with: &Util.embedded_changeset/2)
+    |> cast_embed(:meeting_settings, with: &cast_meeting_settings/2)
     |> cast_embed(:rules, with: &Util.embedded_changeset/2)
+  end
+
+  defp cast_meeting_settings(changeset, params) do
+    changeset
+    |> cast(params, [:video_recording, :audio_recording, :transcription, :action_items, :summary])
+    |> cast_embed(:action_items_settings, with: &Util.embedded_changeset/2)
+    |> cast_embed(:summary_settings, with: &Util.embedded_changeset/2)
   end
 end