+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added new fields to notetaker schema: `action_items`, `action_items_settings`, `summary`, and `summary_settings`
+
 ## [0.10.0] - 2025-08-18
 
 ### Added